Make timeshift_ms a non-optional field
The field timeshift_ms is actually optional and does not necessarily
need to be specified in the schedule. Previously, we used std::option to
describe it, but using std::option makes the code a bit more complex,
and slightly hinders readability when using the SUI RPC API.

Instead of using std::option, I made this field non-optional, with a
default value of zero, which should simplify the code somewhat.