Prune old account history entries
Storing data on the blockchain incurs gas fees, so we should minimize
the data stored. In the vesting ledger, we record the user's account
history to accurately calculate the coins they can claim. This data
grows indefinitely, so we should remove outdated entries.

Once the vesting period is over, we don't need to keep old entries. We
can retain just the last `period` number of entries and delete the rest.

It would be logical to remove old entries in the "claim" function, but
doing so would give the storage refund gas rebate to the user, even
though we paid it on their behalf.

Therefore, instead of pruning old entries in the "claim" function, the
code removes them when new coins are locked in the user's account.